Moved TZ removal to utility function.
