+++ src/applications/check-in/reducers/index.js
@@ -1,3 +1,5 @@
+import { removeTimeZone } from '../utils/appointment';
+
 const initialState = {
   appointments: [],
   context: {},
@@ -32,33 +34,11 @@ const checkInReducer = (state = initialState, action) => {
         context: { ...state.context, ...action.payload.context },
       };
     case RECEIVED_APPOINTMENT_DETAILS: {
-      // Grabing the appointment payload and stripping out timezone here.
-      // Chip should be handling this but currently isn't, this code may be refactored out.
-      const updatedPayload = JSON.parse(JSON.stringify(action.payload));
-      // These fields have a potential to include a time stamp.
-      const timeFields = [
-        'checkInWindowEnd',
-        'checkInWindowStart',
-        'checkedInTime',
-        'startTime',
-      ];
-
-      const updatedAppointments = updatedPayload.appointments.map(
-        appointment => {
-          const updatedAppointment = { ...appointment };
-          // If field exists in object we will replace the TZ part of the string.
-          timeFields.forEach(field => {
-            if (field in updatedAppointment) {
-              updatedAppointment[field] = updatedAppointment[field].replace(
-                /(?=\.).*/,
-                '',
-              );
-            }
-          });
-          return updatedAppointment;
-        },
-      );
-      return { ...state, appointments: updatedAppointments };
+      let payload = JSON.parse(JSON.stringify(action.payload));
+      if ('appointments' in payload) {
+        payload = removeTimeZone(payload);
+      }
+      return { ...state, ...payload };
     }
     case RECEIVED_DEMOGRAPHICS_DATA:
       return { ...state, ...action.payload };
